--- gupb/controller/LordIcon/__init__.py
@@ -1,10 +1,0 @@
-from .lordicon import LordIcon
-
-__all__ = [
-    'LordIcon',
-    'POTENTIAL_CONTROLLERS',
-]
-
-POTENTIAL_CONTROLLERS = [
-    LordIcon("Marko"),
-]
