+++ gupb/controller/lord_icon/distance.py
@@ -1,0 +1,59 @@
+from typing import Tuple, List
+import numpy as np
+import heapq
+
+
+Point2d = Tuple[int, int]
+Direction = Tuple[int, int]
+
+
+def dist(a: Point2d, b: Point2d) -> int:
+    return abs(a[0] - b[0]) + abs(a[1] - b[1])
+
+
+def heuristic(a, b):
+    return np.sqrt((b[0] - a[0]) ** 2 + (b[1] - a[1]) ** 2)
+
+
+def astar(map: np.ndarray, start: Point2d, goal: Point2d) -> List[Point2d]:
+    neighbors = [(0, 1), (0, -1), (1, 0), (-1, 0)]
+    close_set = set()
+    came_from = {}
+    gscore = {start: 0}
+    fscore = {start: heuristic(start, goal)}
+    oheap = []
+    heapq.heappush(oheap, (fscore[start], start))
+
+    while oheap:
+        current = heapq.heappop(oheap)[1]
+        if current == goal:
+            data = []
+            while current in came_from:
+                data.append(current)
+                current = came_from[current]
+            data.reverse()
+            return data
+        close_set.add(current)
+        for i, j in neighbors:
+            neighbor = current[0] + i, current[1] + j
+            tentative_g_score = gscore[current] + heuristic(current, neighbor)
+            if 0 <= neighbor[0] < map.shape[0]:
+                if 0 <= neighbor[1] < map.shape[1]:
+                    if map[neighbor[0]][neighbor[1]] == 1:
+                        continue
+                else:
+                    continue
+            else:
+                continue
+
+            if neighbor in close_set and tentative_g_score >= gscore.get(neighbor, 0):
+                continue
+
+            if tentative_g_score < gscore.get(neighbor, 0) or neighbor not in [
+                i[1] for i in oheap
+            ]:
+                came_from[neighbor] = current
+                gscore[neighbor] = tentative_g_score
+                fscore[neighbor] = tentative_g_score + heuristic(neighbor, goal)
+                heapq.heappush(oheap, (fscore[neighbor], neighbor))
+    return []
